main/requirements.txt
Django==3.2.16
djangorestframework==3.14
psycopg2==2.8.6